pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uroflow-mobile-concept"
version = "0.1.0"
description = "Concept repository for smartphone-video-based uroflowmetry"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "Uroflow R&D Team"}
]
dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
video = [
  "opencv-python>=4.9",
]
clinical-api = [
  "fastapi>=0.116",
  "uvicorn>=0.35",
]
dev = [
  "pytest>=9.0",
  "ruff>=0.6",
  "openpyxl>=3.1",
  "python-docx>=1.1",
  "fastapi>=0.116",
  "uvicorn>=0.35",
  "httpx>=0.28",
]

[project.scripts]
uroflow-mobile = "uroflow_mobile.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
  "scripts/pilot_automation_v2_5/scripts",
  "scripts/pilot_automation_v2_8/scripts",
  "scripts/pilot_automation_v4_2/scripts",
  "tmp",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
